+++ CHANGELOG.md
@@ -7,8 +7,14 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+### Added
+
+* [`Pipeline` and `PipelineHub`](https://laravel.com/docs/10.x/helpers#pipeline) aware-of helpers, in Support package.
+* `PasswordBroker` and `PasswordBrokerManager` aware-of helpers, in Support package.
+
 ### Change
 
+* Minimum required version of Laravel packages changed to `v10.3`.
 * Reduced `sample_size` to 512 bytes, for "file-info" in tests, for the Mime Types package (_Now in accordance with official documentation_).
 * Improved documentation regarding supported types, for the Configuration package.